Create carl and the chocolate factory.py
Carl, the founder of Y-Rubdac Chocolates Inc. has just decided to replace many of his factory workers with machines. This will help him make chocolates at a faster rate and with a better quality. He bought chocolate-machines from a company called Doofenshmirtz Evil Inc.

However, there's a problem with those machines. They do not easily understand what size of chocolates they need to make. What they do understand is a special series of characters called a chocolate string.

A chocolate string is a series of characters "!" and "=". Each chocolate string must start and end with a "!" and must ONLY contain the character "=" between the first and last characters.

The size of the chocolate is equal to the number of "=" characters between the starting and ending "!". For example, "!========!" is a valid chocolate string which represents a chocolate of size 8. Also, "!!===!", "!===!===!", "!!==" are examples of strings which are not chocolate strings.

Carl doesn"t know how this works and needs your help to send the right command to the machine. Given an integer N, you have to output a valid chocolate string that represents a chocoalate of size N.

the above problem's solution is the code.